requirements.txt
arrow==1.3.0
astroid==3.0.2
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
Flask==3.0.1
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
platformdirs==4.2.0
pluggy==1.4.0
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.11.1
pydocstyle==6.3.0
pylint==3.0.3
pyparsing==3.1.4
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
tomlkit==0.12.3
types-python-dateutil==2.8.19.20240106
uritemplate==4.1.1
urllib3==2.2.0
Werkzeug==3.0.1
